adjustment of the price value in the Service model

+++ database/migrations/2025_07_22_161550_create_services_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->string('name');
             $table->string('abbreviation');
-            $table->float('cost');
+            $table->decimal('cost', 5, 2);
             $table->timestamps();
         });
     }
